Refactored editor slightly to make finishing the debugger easier
