+++ crates/bevy_animation_graph_editor/src/ui/core.rs
@@ -1,50 +1,45 @@
 use core::hash::Hasher;
 use std::any::TypeId;
 use std::hash::Hash;
-use std::path::PathBuf;
 
 use crate::asset_saving::{SaveFsm, SaveGraph};
 use crate::egui_fsm::lib::FsmUiContext;
-use crate::egui_inspector_impls::handle_path;
 use crate::egui_nodes::lib::NodesContext;
-use crate::fsm_show::{make_fsm_indices, FsmIndices, FsmReprSpec};
-use crate::graph_show::{make_graph_indices, GraphIndices, GraphReprSpec};
+use crate::fsm_show::{FsmIndices, FsmReprSpec};
+use crate::graph_show::{GraphIndices, GraphReprSpec};
 use crate::graph_update::{
-    apply_global_changes, convert_fsm_change, convert_graph_change, update_graph, Change,
+    apply_global_changes, convert_fsm_change, convert_graph_change, update_graph_asset, Change,
     FsmChange, FsmPropertiesChange, GlobalChange, GraphChange,
 };
 use crate::scanner::PersistedAssetHandles;
-use crate::tree::{Tree, TreeInternal, TreeResult};
+use crate::tree::{Tree, TreeResult};
 use bevy::asset::UntypedAssetId;
 use bevy::ecs::world::CommandQueue;
 use bevy::prelude::*;
-use bevy::render::camera::RenderTarget;
-use bevy::render::render_resource::{
-    Extent3d, TextureDescriptor, TextureDimension, TextureFormat, TextureUsages,
-};
+use bevy::render::render_resource::Extent3d;
 use bevy::utils::HashMap;
 use bevy::window::PrimaryWindow;
-use bevy_animation_graph::core::animated_scene::{
-    AnimatedScene, AnimatedSceneBundle, AnimatedSceneInstance,
-};
+use bevy_animation_graph::core::animated_scene::{AnimatedScene, AnimatedSceneInstance};
 use bevy_animation_graph::core::animation_clip::EntityPath;
-use bevy_animation_graph::core::animation_graph::{AnimationGraph, NodeId, PinMap};
+use bevy_animation_graph::core::animation_graph::{AnimationGraph, NodeId, PinId, SourcePin};
 use bevy_animation_graph::core::animation_graph_player::AnimationGraphPlayer;
 use bevy_animation_graph::core::animation_node::AnimationNode;
-use bevy_animation_graph::core::context::{GraphContext, GraphContextId, SpecContext};
+use bevy_animation_graph::core::context::{GraphContextId, SpecContext};
 use bevy_animation_graph::core::edge_data::AnimationEvent;
 use bevy_animation_graph::core::state_machine::high_level::{
     State, StateId, StateMachine, Transition, TransitionId,
 };
-use bevy_animation_graph::prelude::{
-    AnimatedSceneHandle, DataSpec, ReflectEditProxy, ReflectNodeLike,
-};
+use bevy_animation_graph::prelude::{ReflectEditProxy, ReflectNodeLike};
 use bevy_egui::EguiContext;
 use bevy_inspector_egui::bevy_egui::EguiUserTextures;
 use bevy_inspector_egui::reflect_inspector::{Context, InspectorUi};
 use bevy_inspector_egui::{bevy_egui, egui};
 use egui_dock::{DockArea, DockState, NodeIndex, Style};
 use egui_notify::{Anchor, Toasts};
+
+use super::egui_inspector_impls::handle_path;
+use super::scenes::PreviewScene;
+use super::utils;
 
 pub fn show_ui_system(world: &mut World) {
     let Ok(egui_context) = world
@@ -90,28 +85,29 @@ pub enum InspectorSelection {
 }
 
 pub struct FsmTransitionSelection {
-    fsm: AssetId<StateMachine>,
-    state: TransitionId,
+    pub(crate) fsm: AssetId<StateMachine>,
+    pub(crate) state: TransitionId,
 }
 
 pub struct FsmStateSelection {
-    fsm: AssetId<StateMachine>,
-    state: StateId,
+    pub(crate) fsm: AssetId<StateMachine>,
+    pub(crate) state: StateId,
 }
 
 pub struct NodeSelection {
-    graph: AssetId<AnimationGraph>,
-    node: NodeId,
-    name_buf: String,
+    pub(crate) graph: AssetId<AnimationGraph>,
+    pub(crate) node: NodeId,
+    pub(crate) name_buf: String,
+    pub(crate) selected_pin_id: Option<PinId>,
 }
 
 pub struct SceneSelection {
-    scene: Handle<AnimatedScene>,
-    respawn: bool,
-    active_context: HashMap<UntypedAssetId, GraphContextId>,
-    event_table: Vec<AnimationEvent>,
+    pub(crate) scene: Handle<AnimatedScene>,
+    pub(crate) respawn: bool,
+    pub(crate) active_context: HashMap<UntypedAssetId, GraphContextId>,
+    pub(crate) event_table: Vec<AnimationEvent>,
     /// Just here as a buffer for the editor
-    event_editor: AnimationEvent,
+    pub(crate) event_editor: AnimationEvent,
 }
 
 #[derive(Default)]
@@ -124,9 +120,9 @@ pub struct EditorSelection {
     pub graph_editor: Option<GraphSelection>,
     pub fsm_editor: Option<FsmSelection>,
     pub inspector_selection: InspectorSelection,
-    scene: Option<SceneSelection>,
-    node_creation: NodeCreation,
-    entity_path: Option<EntityPath>,
+    pub(crate) scene: Option<SceneSelection>,
+    pub(crate) node_creation: NodeCreation,
+    pub(crate) entity_path: Option<EntityPath>,
 }
 
 pub enum RequestSave {
@@ -136,26 +132,29 @@ pub enum RequestSave {
 
 #[derive(Resource)]
 pub struct UiState {
-    state: DockState<EguiWindow>,
+    pub(crate) state: DockState<EguiWindow>,
     pub selection: EditorSelection,
-    graph_changes: Vec<GraphChange>,
-    global_changes: Vec<GlobalChange>,
+    pub(crate) graph_changes: Vec<GraphChange>,
+    pub(crate) global_changes: Vec<GlobalChange>,
     /// Requests to save a graph. These still need confirmation from the user,
     /// and specification of path where graph should be saved.
-    save_requests: Vec<RequestSave>,
+    pub(crate) save_requests: Vec<RequestSave>,
     /// Save events to be fired as bevy events after Ui system has finished running
-    graph_save_events: Vec<SaveGraph>,
-    fsm_save_events: Vec<SaveFsm>,
-    preview_image: Handle<Image>,
-    notifications: Toasts,
+    pub(crate) graph_save_events: Vec<SaveGraph>,
+    pub(crate) fsm_save_events: Vec<SaveFsm>,
+    pub(crate) preview_image: Handle<Image>,
+    pub(crate) notifications: Toasts,
 }
 
 impl UiState {
     pub fn new() -> Self {
         let mut state = DockState::new(vec![EguiWindow::GraphEditor, EguiWindow::FsmEditor]);
         let tree = state.main_surface_mut();
-        let [graph_editor, inspectors] =
-            tree.split_right(NodeIndex::root(), 0.75, vec![EguiWindow::Inspector]);
+        let [graph_editor, inspectors] = tree.split_right(
+            NodeIndex::root(),
+            0.75,
+            vec![EguiWindow::Inspector, EguiWindow::Debugger],
+        );
         let [_graph_editor, graph_selector] =
             tree.split_left(graph_editor, 0.2, vec![EguiWindow::GraphSelector]);
         let [_graph_selector, scene_selector] =
@@ -191,7 +190,7 @@ impl UiState {
     fn ui(&mut self, world: &mut World, ctx: &mut egui::Context) {
         for save_request in self.save_requests.drain(..) {
             self.state
-                .add_window(vec![TabViewer::create_saver_window(world, save_request)]);
+                .add_window(vec![utils::create_saver_window(world, save_request)]);
         }
         let mut tab_viewer = TabViewer {
             world,
@@ -213,7 +212,7 @@ impl UiState {
 }
 
 #[derive(Debug)]
-enum EguiWindow {
+pub enum EguiWindow {
     GraphEditor,
     FsmEditor,
     Preview,
@@ -329,7 +328,11 @@ impl egui_dock::TabViewer for TabViewer<'_> {
             let must_regen_indices = self.world.resource_scope::<Assets<AnimationGraph>, _>(
                 |world, mut graph_assets| {
                     world.resource_scope::<Assets<StateMachine>, _>(|_, fsm_assets| {
-                        update_graph(self.graph_changes.clone(), &mut graph_assets, &fsm_assets)
+                        update_graph_asset(
+                            self.graph_changes.clone(),
+                            &mut graph_assets,
+                            &fsm_assets,
+                        )
                     })
                 },
             );
@@ -337,7 +340,7 @@ impl egui_dock::TabViewer for TabViewer<'_> {
             if must_regen_indices {
                 if let Some(graph_selection) = self.selection.graph_editor.as_mut() {
                     graph_selection.graph_indices =
-                        Self::update_graph_indices(self.world, graph_selection.graph);
+                        utils::update_graph_indices(self.world, graph_selection.graph);
                 }
             }
         }
@@ -346,11 +349,11 @@ impl egui_dock::TabViewer for TabViewer<'_> {
         if must_regen_indices {
             if let Some(graph_selection) = self.selection.graph_editor.as_mut() {
                 graph_selection.graph_indices =
-                    Self::update_graph_indices(self.world, graph_selection.graph);
+                    utils::update_graph_indices(self.world, graph_selection.graph);
             }
             if let Some(fsm_selection) = self.selection.fsm_editor.as_mut() {
                 fsm_selection.graph_indices =
-                    Self::update_fsm_indices(self.world, fsm_selection.fsm);
+                    utils::update_fsm_indices(self.world, fsm_selection.fsm);
             }
         }
         self.global_changes.clear();
@@ -432,7 +435,7 @@ impl TabViewer<'_> {
         };
         let entity = instance.player_entity;
         let tree = Tree::entity_tree(world, entity);
-        match Self::select_from_branches(ui, tree.0) {
+        match utils::select_from_branches(ui, tree.0) {
             TreeResult::Leaf((_, path)) => {
                 let path = EntityPath { parts: path };
                 ui.output_mut(|o| o.copied_text = path.to_slashed_string());
@@ -477,7 +480,7 @@ impl TabViewer<'_> {
                     // Autoselect context if none selected and some available
                     if let (Some(scene), Some(available_contexts)) = (
                         &mut selection.scene,
-                        list_graph_contexts(world, |ctx| {
+                        utils::list_graph_contexts(world, |ctx| {
                             ctx.get_graph_id() == graph_selection.graph
                         }),
                     ) {
@@ -493,7 +496,7 @@ impl TabViewer<'_> {
                         }
                     }
 
-                    let graph_player = get_animation_graph_player(world);
+                    let graph_player = utils::get_animation_graph_player(world);
 
                     let maybe_graph_context = selection
                         .scene
@@ -559,6 +562,7 @@ impl TabViewer<'_> {
                                     graph: graph_selection.graph,
                                     node: node_name.clone(),
                                     name_buf: node_name.clone(),
+                                    selected_pin_id: None,
                                 });
                         }
                     }
@@ -599,7 +603,7 @@ impl TabViewer<'_> {
             return;
         };
         let Some(graph_player) =
-            get_animation_graph_player_mut(unsafe { unsafe_world.world_mut() })
+            utils::get_animation_graph_player_mut(unsafe { unsafe_world.world_mut() })
         else {
             return;
         };
@@ -655,7 +659,7 @@ impl TabViewer<'_> {
                     // Autoselect context if none selected and some available
                     if let (Some(scene), Some(available_contexts)) = (
                         &mut selection.scene,
-                        list_graph_contexts(world, |ctx| {
+                        utils::list_graph_contexts(world, |ctx| {
                             let graph_id = ctx.get_graph_id();
                             graph_assets
                                 .get(graph_id)
@@ -675,7 +679,7 @@ impl TabViewer<'_> {
                         }
                     }
 
-                    let graph_player = get_animation_graph_player(world);
+                    let graph_player = utils::get_animation_graph_player(world);
 
                     let maybe_fsm_state = selection
                         .scene
@@ -780,7 +784,7 @@ impl TabViewer<'_> {
                     .into_iter()
                     .map(|id| (handle_path(id.untyped(), &asset_server), id))
                     .collect();
-                if let TreeResult::Leaf(id) = Self::path_selector(ui, paths) {
+                if let TreeResult::Leaf(id) = utils::path_selector(ui, paths) {
                     chosen_id = Some(id);
                 }
                 ui.with_layout(egui::Layout::bottom_up(egui::Align::Center), |ui| {
@@ -798,7 +802,7 @@ impl TabViewer<'_> {
         if let Some(chosen_id) = chosen_id {
             selection.graph_editor = Some(GraphSelection {
                 graph: chosen_id,
-                graph_indices: Self::update_graph_indices(world, chosen_id),
+                graph_indices: utils::update_graph_indices(world, chosen_id),
                 nodes_context: NodesContext::default(),
             });
             selection.inspector_selection = InspectorSelection::Graph;
@@ -819,7 +823,7 @@ impl TabViewer<'_> {
                     .into_iter()
                     .map(|id| (handle_path(id.untyped(), &asset_server), id))
                     .collect();
-                let chosen_id = Self::path_selector(ui, paths);
+                let chosen_id = utils::path_selector(ui, paths);
                 if let TreeResult::Leaf(id) = chosen_id {
                     chosen_handle = Some(
                         asset_server
@@ -862,7 +866,7 @@ impl TabViewer<'_> {
                     .into_iter()
                     .map(|id| (handle_path(id.untyped(), &asset_server), id))
                     .collect();
-                if let TreeResult::Leaf(id) = Self::path_selector(ui, paths) {
+                if let TreeResult::Leaf(id) = utils::path_selector(ui, paths) {
                     chosen_id = Some(id);
                 }
                 // ui.with_layout(egui::Layout::bottom_up(egui::Align::Center), |ui| {
@@ -875,7 +879,7 @@ impl TabViewer<'_> {
         if let Some(chosen_id) = chosen_id {
             selection.fsm_editor = Some(FsmSelection {
                 fsm: chosen_id,
-                graph_indices: Self::update_fsm_indices(world, chosen_id),
+                graph_indices: utils::update_fsm_indices(world, chosen_id),
                 nodes_context: FsmUiContext::default(),
                 state_creation: State::default(),
                 transition_creation: Transition::default(),
@@ -892,7 +896,7 @@ impl TabViewer<'_> {
     ) {
         ui.heading("Animation graph");
 
-        select_graph_context(world, ui, selection);
+        utils::select_graph_context(world, ui, selection);
 
         ui.collapsing("Create node", |ui| {
             Self::node_creator(world, ui, selection, graph_changes)
@@ -951,7 +955,7 @@ impl TabViewer<'_> {
         ui.heading("State machine");
         let mut changes = Vec::new();
 
-        select_graph_context_fsm(world, ui, selection);
+        utils::select_graph_context_fsm(world, ui, selection);
 
         let Some(fsm_selection) = &mut selection.fsm_editor else {
             return;
@@ -1509,6 +1513,35 @@ impl TabViewer<'_> {
         };
         let entity = instance.player_entity;
         let mut query = world.query::<&AnimationGraphPlayer>();
+
+        let unsafe_world = world.as_unsafe_world_cell();
+        let type_registry = unsafe {
+            unsafe_world
+                .get_resource::<AppTypeRegistry>()
+                .unwrap()
+                .0
+                .clone()
+        };
+        let type_registry = type_registry.read();
+        let mut queue = CommandQueue::default();
+        let mut cx = Context {
+            world: Some(unsafe { unsafe_world.world_mut() }.into()),
+            queue: Some(&mut queue),
+        };
+        let mut env = InspectorUi::for_bevy(&type_registry, &mut cx);
+
+        let world = unsafe { unsafe_world.world_mut() };
+
+        let InspectorSelection::Node(node_selection) = &mut selection.inspector_selection else {
+            return;
+        };
+
+        let Some(data_pin_map) =
+            utils::get_node_output_data_pins(world, node_selection.graph, &node_selection.node)
+        else {
+            return;
+        };
+
         let Ok(player) = query.get(world, entity) else {
             return;
         };
@@ -1527,413 +1560,43 @@ impl TabViewer<'_> {
             return;
         };
 
-        let InspectorSelection::Node(node_selection) = &selection.inspector_selection else {
-            return;
-        };
-
-        let Some(data_pin_map) =
-            Self::get_node_output_data_pins(world, node_selection.graph, &node_selection.node)
-        else {
-            return;
-        };
-
         // Select data to display cached value of
-        let curr_val = "".to_string();
-        let mut new_val = curr_val.clone();
+        let mut selected_pin_id = node_selection
+            .selected_pin_id
+            .clone()
+            .unwrap_or("".to_string());
         egui::ComboBox::new("cache debugger window", "Pin")
-            .selected_text(curr_val)
+            .selected_text(&selected_pin_id)
             .show_ui(ui, |ui| {
                 for (id, spec) in data_pin_map.iter() {
-                    ui.selectable_value(&mut new_val, id.clone(), id)
+                    ui.selectable_value(&mut selected_pin_id, id.clone(), id)
                         .on_hover_ui(|ui| {
                             ui.label(format!("{:?}", spec));
                         });
                 }
             });
+        let new_pin_id = if selected_pin_id == "" {
+            None
+        } else {
+            Some(selected_pin_id)
+        };
+
+        if new_pin_id != node_selection.selected_pin_id {
+            node_selection.selected_pin_id = new_pin_id.clone();
+        }
 
         // Now get the selected value and display it!
 
-        // graph_context.caches.get_primary(|s|);
-    }
-}
-
-/// Helper functions
-impl TabViewer<'_> {
-    fn get_node_output_data_pins(
-        world: &mut World,
-        graph_id: AssetId<AnimationGraph>,
-        node_id: &NodeId,
-    ) -> Option<PinMap<DataSpec>> {
-        world.resource_scope::<Assets<AnimationGraph>, _>(|world, graph_assets| {
-            world.resource_scope::<Assets<StateMachine>, _>(|_, fsm_assets| {
-                let graph = graph_assets.get(graph_id).unwrap();
-                let spec_context = SpecContext {
-                    graph_assets: &graph_assets,
-                    fsm_assets: &fsm_assets,
-                };
-                let node = graph.nodes.get(node_id)?;
-                Some(node.inner.data_output_spec(spec_context))
+        let Some(val) = new_pin_id.and_then(|pin_id| {
+            graph_context.caches.get_primary(|c| {
+                let node_id = node_selection.node.clone();
+                let pin_id = pin_id.clone();
+                c.get_data(&SourcePin::NodeData(node_id, pin_id)).cloned()
             })
-        })
+        }) else {
+            return;
+        };
+
+        env.ui_for_reflect_readonly(&val, ui);
     }
-
-    fn create_saver_window(world: &mut World, save_request: RequestSave) -> EguiWindow {
-        world.resource_scope::<AssetServer, EguiWindow>(|_, asset_server| match save_request {
-            RequestSave::Graph(graph) => {
-                let path = asset_server
-                    .get_path(graph)
-                    .map_or("".into(), |p| p.path().to_string_lossy().into());
-                EguiWindow::GraphSaver(graph, path, false)
-            }
-            RequestSave::Fsm(fsm_id) => {
-                let path = asset_server
-                    .get_path(fsm_id)
-                    .map_or("".into(), |p| p.path().to_string_lossy().into());
-                EguiWindow::FsmSaver(fsm_id, path, false)
-            }
-        })
-    }
-
-    fn path_selector<T>(ui: &mut egui::Ui, paths: Vec<(PathBuf, T)>) -> TreeResult<(), T> {
-        // First, preprocess paths into a tree structure
-        let mut tree = Tree::default();
-        for (path, val) in paths {
-            let parts: Vec<String> = path
-                .components()
-                .map(|c| c.as_os_str().to_string_lossy().into())
-                .collect();
-            tree.insert(parts, val);
-        }
-
-        // Then, display the tree
-        Self::select_from_branches(ui, tree.0)
-    }
-
-    fn select_from_branches<I, L>(
-        ui: &mut egui::Ui,
-        branches: Vec<TreeInternal<I, L>>,
-    ) -> TreeResult<I, L> {
-        let mut res = TreeResult::None;
-
-        for branch in branches {
-            res = res.or(Self::select_from_tree_internal(ui, branch));
-        }
-
-        res
-    }
-
-    fn select_from_tree_internal<I, L>(
-        ui: &mut egui::Ui,
-        tree: TreeInternal<I, L>,
-    ) -> TreeResult<I, L> {
-        match tree {
-            TreeInternal::Leaf(name, val) => {
-                if ui.selectable_label(false, name).clicked() {
-                    TreeResult::Leaf(val)
-                } else {
-                    TreeResult::None
-                }
-            }
-            TreeInternal::Node(name, val, subtree) => {
-                let res = ui.collapsing(name, |ui| Self::select_from_branches(ui, subtree));
-                if res.header_response.clicked() {
-                    TreeResult::Node(val)
-                } else {
-                    TreeResult::None
-                }
-                .or(res.body_returned.unwrap_or(TreeResult::None))
-                //.body_returned
-                //.flatten(),
-            }
-        }
-    }
-
-    fn update_graph(world: &mut World, changes: Vec<GraphChange>) -> bool {
-        world.resource_scope::<Assets<AnimationGraph>, _>(|world, mut graph_assets| {
-            world.resource_scope::<Assets<StateMachine>, _>(|_, fsm_assets| {
-                update_graph(changes, &mut graph_assets, &fsm_assets)
-            })
-        })
-    }
-
-    fn update_graph_indices(world: &mut World, graph_id: AssetId<AnimationGraph>) -> GraphIndices {
-        let mut res = Self::indices_one_step(world, graph_id);
-
-        while let Err(changes) = &res {
-            Self::update_graph(world, changes.clone());
-            res = Self::indices_one_step(world, graph_id);
-        }
-
-        res.unwrap()
-    }
-
-    fn update_fsm_indices(world: &mut World, fsm_id: AssetId<StateMachine>) -> FsmIndices {
-        world.resource_scope::<Assets<StateMachine>, FsmIndices>(|_, fsm_assets| {
-            let fsm = fsm_assets.get(fsm_id).unwrap();
-
-            make_fsm_indices(fsm, &fsm_assets).unwrap()
-        })
-    }
-
-    fn indices_one_step(
-        world: &mut World,
-        graph_id: AssetId<AnimationGraph>,
-    ) -> Result<GraphIndices, Vec<GraphChange>> {
-        world.resource_scope::<Assets<AnimationGraph>, _>(|world, graph_assets| {
-            world.resource_scope::<Assets<StateMachine>, _>(|_, fsm_assets| {
-                let graph = graph_assets.get(graph_id).unwrap();
-                let spec_context = SpecContext {
-                    graph_assets: &graph_assets,
-                    fsm_assets: &fsm_assets,
-                };
-
-                match make_graph_indices(graph, spec_context) {
-                    Err(targets) => Err(targets
-                        .into_iter()
-                        .map(|t| GraphChange {
-                            graph: graph_id,
-                            change: Change::LinkRemoved(t),
-                        })
-                        .collect()),
-                    Ok(indices) => Ok(indices),
-                }
-            })
-        })
-    }
-}
-
-fn list_graph_contexts(
-    world: &mut World,
-    filter: impl Fn(&GraphContext) -> bool,
-) -> Option<Vec<GraphContextId>> {
-    let player = get_animation_graph_player(world)?;
-    let arena = player.get_context_arena()?;
-
-    Some(
-        arena
-            .iter_context_ids()
-            .filter(|id| {
-                let context = arena.get_context(*id).unwrap();
-                filter(context)
-            })
-            .collect(),
-    )
-}
-
-fn select_graph_context(world: &mut World, ui: &mut egui::Ui, selection: &mut EditorSelection) {
-    let Some(graph) = &selection.graph_editor else {
-        return;
-    };
-
-    let Some(available) = list_graph_contexts(world, |ctx| ctx.get_graph_id() == graph.graph)
-    else {
-        return;
-    };
-
-    let Some(scene) = &mut selection.scene else {
-        return;
-    };
-
-    let mut selected = scene.active_context.get(&graph.graph.untyped()).copied();
-    egui::ComboBox::from_label("Active context")
-        .selected_text(format!("{:?}", selected))
-        .show_ui(ui, |ui| {
-            ui.selectable_value(&mut selected, None, format!("{:?}", None::<GraphContextId>));
-            for id in available {
-                ui.selectable_value(&mut selected, Some(id), format!("{:?}", Some(id)));
-            }
-        });
-
-    if let Some(selected) = selected {
-        scene.active_context.insert(graph.graph.untyped(), selected);
-    } else {
-        scene.active_context.remove(&graph.graph.untyped());
-    }
-}
-
-fn select_graph_context_fsm(world: &mut World, ui: &mut egui::Ui, selection: &mut EditorSelection) {
-    let Some(fsm) = &selection.fsm_editor else {
-        return;
-    };
-
-    let Some(available) =
-        world.resource_scope::<Assets<AnimationGraph>, _>(|world, graph_assets| {
-            list_graph_contexts(world, |ctx| {
-                let graph_id = ctx.get_graph_id();
-                let graph = graph_assets.get(graph_id).unwrap();
-                graph.contains_state_machine(fsm.fsm).is_some()
-            })
-        })
-    else {
-        return;
-    };
-
-    let Some(scene) = &mut selection.scene else {
-        return;
-    };
-
-    let mut selected = scene.active_context.get(&fsm.fsm.untyped()).copied();
-    egui::ComboBox::from_label("Active context")
-        .selected_text(format!("{:?}", selected))
-        .show_ui(ui, |ui| {
-            ui.selectable_value(&mut selected, None, format!("{:?}", None::<GraphContextId>));
-            for id in available {
-                ui.selectable_value(&mut selected, Some(id), format!("{:?}", Some(id)));
-            }
-        });
-
-    if let Some(selected) = selected {
-        scene.active_context.insert(fsm.fsm.untyped(), selected);
-    } else {
-        scene.active_context.remove(&fsm.fsm.untyped());
-    }
-}
-
-#[derive(Component)]
-pub struct PreviewScene;
-
-pub fn scene_spawner_system(
-    mut commands: Commands,
-    mut query: Query<(Entity, &AnimatedSceneHandle), With<PreviewScene>>,
-    mut ui_state: ResMut<UiState>,
-) {
-    if let Ok((entity, AnimatedSceneHandle(scene_handle))) = query.get_single_mut() {
-        if let Some(scene_selection) = &mut ui_state.selection.scene {
-            if scene_selection.respawn || &scene_selection.scene != scene_handle {
-                commands.entity(entity).despawn_recursive();
-                commands
-                    .spawn(AnimatedSceneBundle {
-                        animated_scene: AnimatedSceneHandle(scene_selection.scene.clone()),
-                        ..default()
-                    })
-                    .insert(PreviewScene);
-                scene_selection.respawn = false;
-            }
-        } else {
-            commands.entity(entity).despawn_recursive();
-        }
-    } else if let Some(scene_selection) = &mut ui_state.selection.scene {
-        commands
-            .spawn(AnimatedSceneBundle {
-                animated_scene: AnimatedSceneHandle(scene_selection.scene.clone()),
-                ..default()
-            })
-            .insert(PreviewScene);
-        scene_selection.respawn = false;
-    }
-}
-
-pub fn asset_save_event_system(
-    mut ui_state: ResMut<UiState>,
-    mut evw_save_graph: EventWriter<SaveGraph>,
-    mut evw_save_fsm: EventWriter<SaveFsm>,
-) {
-    for save_event in ui_state.graph_save_events.drain(..) {
-        evw_save_graph.send(save_event);
-    }
-    for save_event in ui_state.fsm_save_events.drain(..) {
-        evw_save_fsm.send(save_event);
-    }
-}
-
-pub fn graph_debug_draw_bone_system(
-    ui_state: Res<UiState>,
-    scene_instance_query: Query<&AnimatedSceneInstance, With<PreviewScene>>,
-    mut player_query: Query<&mut AnimationGraphPlayer>,
-) {
-    let Some(path) = ui_state.selection.entity_path.as_ref() else {
-        return;
-    };
-    if ui_state.selection.scene.is_none() {
-        return;
-    };
-    let Ok(instance) = scene_instance_query.get_single() else {
-        return;
-    };
-    let entity = instance.player_entity;
-    let Ok(mut player) = player_query.get_mut(entity) else {
-        return;
-    };
-
-    player.gizmo_for_bones(vec![path.clone().id()])
-}
-
-pub fn setup_system(
-    mut egui_user_textures: ResMut<bevy_egui::EguiUserTextures>,
-    mut ui_state: ResMut<UiState>,
-    mut commands: Commands,
-    mut images: ResMut<Assets<Image>>,
-) {
-    let size = Extent3d {
-        width: 512,
-        height: 512,
-        ..default()
-    };
-
-    // This is the texture that will be rendered to.
-    let mut image = Image {
-        texture_descriptor: TextureDescriptor {
-            label: None,
-            size,
-            dimension: TextureDimension::D2,
-            format: TextureFormat::Bgra8UnormSrgb,
-            mip_level_count: 1,
-            sample_count: 1,
-            usage: TextureUsages::TEXTURE_BINDING
-                | TextureUsages::COPY_DST
-                | TextureUsages::RENDER_ATTACHMENT,
-            view_formats: &[],
-        },
-        ..default()
-    };
-
-    // fill image.data with zeroes
-    image.resize(size);
-
-    let image_handle = images.add(image);
-
-    egui_user_textures.add_image(image_handle.clone());
-    ui_state.preview_image = image_handle.clone();
-
-    // Light
-    // NOTE: Currently lights are shared between passes - see https://github.com/bevyengine/bevy/issues/3462
-    commands.spawn((
-        PointLight::default(),
-        Transform::from_translation(Vec3::new(0.0, 0.0, 10.0)),
-    ));
-
-    commands.spawn((
-        Camera3d::default(),
-        Camera {
-            // render before the "main pass" camera
-            order: -1,
-            clear_color: ClearColorConfig::Custom(Color::from(LinearRgba::new(1.0, 1.0, 1.0, 0.0))),
-            target: RenderTarget::Image(image_handle),
-            ..default()
-        },
-        Transform::from_translation(Vec3::new(0.0, 2.0, 3.0)).looking_at(Vec3::Y, Vec3::Y),
-    ));
-}
-
-fn get_animation_graph_player(world: &mut World) -> Option<&AnimationGraphPlayer> {
-    let mut query = world.query::<(&AnimatedSceneInstance, &PreviewScene)>();
-    let Ok((instance, _)) = query.get_single(world) else {
-        return None;
-    };
-    let entity = instance.player_entity;
-    let mut query = world.query::<&AnimationGraphPlayer>();
-    query.get(world, entity).ok()
-}
-
-fn get_animation_graph_player_mut(world: &mut World) -> Option<&mut AnimationGraphPlayer> {
-    let mut query = world.query::<(&AnimatedSceneInstance, &PreviewScene)>();
-    let Ok((instance, _)) = query.get_single(world) else {
-        return None;
-    };
-    let entity = instance.player_entity;
-    let mut query = world.query::<&mut AnimationGraphPlayer>();
-    query
-        .get_mut(world, entity)
-        .ok()
-        .map(|player| player.into_inner())
 }
